pyproject.toml
[project]
name = "practice-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=1.6.1",
    "dotenv>=0.9.9",
    "fpdf>=1.7.2",
    "google-generativeai>=0.8.6",
    "kaggle>=2.0.0",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.1",
    "numpy>=2.4.3",
    "pdf2image>=1.17.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.55.0",
]
